Rename `base.py` to `base_client.py`
