+++ jardin/database/clients/mysql.py
@@ -1,6 +1,6 @@
 import pymysql
 
-from jardin.database.base import BaseClient
+from jardin.database.base_client import BaseClient
 from jardin.database.lexicon import BaseLexicon
