+++ jardin/database/clients/sqlite.py
@@ -1,6 +1,6 @@
 import sqlite3
 
-from jardin.database.base import BaseClient
+from jardin.database.base_client import BaseClient
 from jardin.database.lexicon import BaseLexicon
